requirements.txt
streamlit==1.20.0
requests==2.28.1
python-dotenv==1.0.0
